Fix BYOK Jest imports and serializable fallback
