+++ apps/web/src/app/(app)/issues/page.tsx
@@ -9,7 +9,7 @@ import { parseAsJson, useQueryState } from "nuqs";
 import { useAuth } from "src/core/providers/auth.provider";
 import { usePermissions } from "src/core/providers/permissions.provider";
 import { filterArray, type FilterValueGroup } from "src/core/utils/filtering";
-import { hasPermission } from "src/core/utils/permissions";
+import { hasPermission } from "src/core/utils/permission-map";
 import { safeArray } from "src/core/utils/safe-array";
 
 import { IssuesDataTable } from "./_components/data-table";
