+++ apps/web/src/core/utils/permissions.ts
@@ -1,11 +1,8 @@
 import { NextRequest, NextResponse } from "next/server";
 import { UserRole } from "@enums";
-import {
-    Action,
-    PermissionsMap,
-    ResourceType,
-} from "@services/permissions/types";
+import { ResourceType } from "@services/permissions/types";
 import type { Session } from "next-auth";
+import { hasPermission } from "./permission-map";
 
 const resourceRoutes = {
     [ResourceType.All]: [
@@ -132,35 +129,4 @@ export function handleAuthenticated(
     return next;
 }
 
-export const hasPermission = (params: {
-    permissions: PermissionsMap;
-    organizationId: string;
-    action: Action;
-    resource: ResourceType;
-    repoId?: string;
-}) => {
-    const { permissions, organizationId, action, resource, repoId } = params;
-
-    if (!permissions || !organizationId) {
-        return false;
-    }
-
-    if (permissions[ResourceType.All]?.[Action.Manage]) {
-        return true;
-    }
-
-    const resourcePermissions = permissions[resource]?.[action];
-
-    if (!resourcePermissions) {
-        return false;
-    }
-
-    const matchOrgId = resourcePermissions.organizationId === organizationId;
-
-    let matchRepoId = true;
-    if (repoId && resourcePermissions.repoId) {
-        matchRepoId = resourcePermissions.repoId.includes(repoId);
-    }
-
-    return matchOrgId && matchRepoId;
-};
+export { hasPermission };
